refactor(semantic_visitor): comment out unique_type_visitor
For now it is up to the user to not add conflicting types. I don't know how conflicting types would be handled as we only deem rng and seq conflicting because they both increment the prompt index but there are more events than incrementing. Should this be manually handled by the author? They shouldn't know of the implementation (incrementing pointers)..